chore(example): Fix broken links in docusaurus config
The example site build was failing due to broken links pointing to a non-existent /blog page. These links were present in the default navbar and footer configuration.

This commit resolves the build error by removing the blog preset and all links to the blog from the docusaurus.config.ts file.