feat(plugin): Decouple plugin ID to support multiple instances
Refactored the plugin architecture to remove a hardcoded plugin ID.

Previously, the `ResistanceTable` component assumed the plugin instance ID was always 'example-resistogram', making the plugin non-reusable and preventing the use of multiple plugin instances in the same Docusaurus site.

This commit introduces the following changes:
- The remark plugin now accepts a `pluginId` option.
- This `pluginId` is passed as a prop to the `<ResistanceTable>` component during MDX processing.
- The `ResistanceTable` component uses this `pluginId` prop to dynamically look up its corresponding data from the global plugin data object.
- The example `docusaurus.config.ts` has been updated to demonstrate the correct configuration.

This change makes the plugin robustly reusable and allows users to configure multiple, independent instances of the resistogram plugin.