feat: Make data download script more robust
The data download script is now more robust. Instead of relying on a fixed list of files, it now dynamically determines the files to download based on the content of the data_sources.csv file.

This change makes the script more resilient to changes in the upstream data source and ensures that the example data is always up-to-date.